requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-form==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
requests==2.32.3
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
